moved log/profiler to h/
